WIP Fixing Siri updating widgets
